backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
websocket-client
Pillow